+++ CHANGES.rst
@@ -84,6 +84,7 @@ Development
 - Object model refactoring
 - HTTP API: Add endpoint for getting the last reading
 - Refactor ``sensors``-section of configuration settings
+- Settings: Rename sensor "key" attribute to sensor "id"
 
 
 2019-06-22 0.5.1